tools/EPLLM/requirements.txt
# EPLLM standalone (no dependency on tools/pllm)
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
docarray>=0.16.0
# Core dependencies (same as pllm for PyPI/Docker/LLM)
pypi-json>=0.4.0
docker>=7.0.0
requests>=2.28.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
